Scene detect changes (#383)
* Remove sc-downscale-height CLI option

The standard scenecut detection method at 1080p
is currently as fast as downsampling to 720p
and running scenecut detection on that, and the
fast scenecut method is significantly faster
without downsampling, because the algorithm
is faster than ffmpeg's resizing algorithm.

As such, it doesn't make sense to have this option anymore.

* Fix conditional use of newer VS APIs

It makes sense to enable these by default,
since the vast majority of users
should have a moderately recent version
of Vapoursynth.

We also needed to disable the static ffmpeg
as the default for now. libaom returns None
for the pixel format, so it is not usable
for the changes needed later in this PR.
I attempted to get the git version of
ffmpeg-next compiling with libdav1d,
but was running into issues, I believe
because Arch's package manager only includes
a dynamic library for libdav1d and not
a static version. It probably makes sense
to disable it anyway given the issues
Windows users have been having with it.

* No longer convert input to 8-bit for scene detect

It appears that downconverting does have an impact
on scene detection accuracy. The speed penalty
for running in 10-bit compared to 8-bit is only
10%. Given this, it seems preferable to run
the scene detect in the more accurate format.

* Skip ffmpeg intermediate step for vapoursynth input

* Fix skipping of ffmpeg pipe step

This optimization added in #379 was never triggered.
Because the initialize() method always added a set
of default params to self.ffmpeg_params, the clause
to skip the ffmpeg pipe, self.ffmpeg_params.is_empty(),
was unreachable.

* Revert "Remove sc-downscale-height CLI option"

This reverts commit a51763607faddae07a595f567d4f4cf815de3299.

* Validate pixel formats based on encoder support

* Fix CI build

The dynamic linking of ffmpeg doesn't seem to play nicely with
the old version of ffmpeg which is packaged in Ubuntu's last LTS.
This enables static linking when we are running in CI to work around
that issue, given that that workaround is easier than overhauling
which docker container we use.