Automatically add cq-level arg if it is not present (#460)
* Automatically add cq-level arg if it is not present

Currently, target quality requires a cq-level (or equivalent for the
encoder) arg to be present in order to work, so that it can replace the
existing cq-level with the new one. This commit makes it so that if the
arg is not present, av1an will insert it with the proper quality level.

* Use format! for insert_q

Co-authored-by: Redzic <yusufredzic@gmail.com>