Do not automatically backport each PR to stable
The way we have been doing releases does not coincide well with how we
have been utilizing our stable branch. Given that we release from
master, and that many people in the community use master on a daily
basis, it makes more sense for stable to always match the latest
released version, and for master to include unreleased changes which are
considered "stable".

For experiments, we should instead use a feature branch per experiment
until we are confident that it is ready to be merged into master.