Perform scene detection in parallel with encoding
This (very experimental) commit performs scene detection in parallel
with encoding, to bypass the threading limitations of scene detection.
As new scenes are detected, they are added to the chunk queue and
encoding is immediately started.

There are a number of issues with this initial implementation. To name
a few, this implementation only works with vapoursynth chunking, and
it breaks mkvmerge concatenation and resuming.