Add additional flags, options (#671)
* Add never overwrite -n flag

Acts as the opposite of the overwrite -y flag

* Add --scaler flag

This allows the user to control which scaling algorithm is used during scene detection downscaling and VMAF calculations, and also allows the width of the lanczos scaler to be selected from 1 to 9

* Add inputres option for --vmaf-res

This allows the user to have VMAF calculations use the input resolution automatically, without having to type it in manually per file

* Change progress bar characters for Windows

This looks smoother on the default Windows command prompt

* Add --extra-splits-sec option

Allows the user to specify extra splits in seconds. If both frames and seconds are specified, frames will take priority

* Add --ignore-frame-mismatch flag

Allows the user to ignore any reported frame mismatches between the encoder and chunk frame counts, which is useful if an ffmpeg filter changes the frame count, or the input video is encoded badly (ex. "Missing key frame while searching for timestamp" ffmpeg warning)

* formatting

* Fix compilation

oops

---------

Co-authored-by: Josh Holmer <jholmer.in@gmail.com>