Fix clippy issues and stop it from breaking with every Rust update
The lints in `nursery` and `pedantic` are not intended to be stable,
and many of them will be added or changed with each Rust update.
This makes it annoying to maintain the package, because CI will
begin failing every 6 weeks when a new Rust version releases.
This cleans up this version's new pedantic lints, and then removes
our overrides that enable the unstable lints.

It looks like this was already done in the `av1an` package a while ago,
but we never did it for `av1an-core`.