Add CTRL+C SIGINT handling
Sending SIGINT (pressing CTRL+C) while encoding currently terminates the chunk encoder, causing Av1an to retry the chunk and requiring the user to send multiple signals to terminate. This PR aims to bring that down to 1 command. There are also 2 open issues requesting that sending SIGINT should let the chunk(s) finish but not start any more until all are completed and Av1an exits early with no progress lost. That is the secondary goal for this PR but I'm not sure how to implement that yet.