Set the output FPS when concatenating
This ensures that certain encoders such as vpx which do not properly set
the frame rate on their outputs do not cause av1an to output videos with
the wrong frame rate. This only works with mkvmerge concat method.
As a result of this plus the existing issue with ffmpeg not
concatenating x265 outputs correctly, this also changes the default
concat method to mkvmerge.

Fixes #475